+++ CHANGELOG.md
@@ -19,7 +19,7 @@ The format is based on [Keep a Changelog][kac], and this project adheres to
 
 ### Security
 
-## TBR — 2026-06-16
+## 26.06.0 — 2026-06-16
 
 ### Added